+++ mimesis/data/int/__init__.py
@@ -16,4 +16,3 @@ from .scientific import *
 from .structure import *
 from .text import *
 from .transport import *
-from .games import *